+++ packages/adminui-vue/utils/Form/processing.js
@@ -60,17 +60,18 @@ const {
  * @param {number} instanceID instanceID
  * @param {Object} [parentContext] Optional values for main instance attributes passed to addNew method
  * @param {UBEntity} entitySchema Entity schema
- * @param {function} [beforeInit] Callback which will be emit before init
- * @param {function} [inited] Callback which will be emit when data is inited
- * @param {function} [beforeSave] Callback which will be emit before save
- * @param {function} [saved] Callback which will be emit when data was saved
- * @param {function} [beforeDelete] Callback which will be emit before delete
- * @param {function} [deleted] Callback which will be emit when data was deleted
- * @param {function} [beforeCopy] Callback which will be emit before copy of existing record
- * @param {function} [copied] Callback which will be emit when data was copied from existing record
- * @param {function} [saveNotification] Callback which will be override default save notification
- * @param {boolean} [isCopy] Flag which used for create new record with data of existing record
- * @param {boolean} [isModal] Is parent opened from modal. Used to provide modal state to child
+ * @param {function} [beforeInit] Callback that emits before init
+ * @param {function} [inited] Callback that emits when data is inited
+ * @param {function} [beforeSave] Callback that emits before save
+ * @param {function} [saved] Callback that emits when data was saved
+ * @param {function} [beforeDelete] Callback that emits before delete
+ * @param {function} [deleted] Callback that emits when data was deleted
+ * @param {function} [beforeCopy] Callback that emits before copy of existing record
+ * @param {function} [copied] Callback that emits when data was copied from existing record
+ * @param {function} [saveNotification] Callback that overrides default save notification
+ * @param {function} [errorNotification] Callback that overrides default error notification
+ * @param {boolean} [isCopy] Flag that used for create new record with data of existing record
+ * @param {boolean} [isModal] Is parent opened from modal. Used to provide modal state to the child
  * @return {object} Vue store cfg
  */
 function createProcessingModule ({
@@ -94,6 +95,7 @@ function createProcessingModule ({
   beforeCopy,
   copied,
   saveNotification,
+  errorNotification,
   isCopy,
   isModal
 }) {
@@ -337,7 +339,7 @@ function createProcessingModule ({
       },
 
       /**
-       * Add a new item to a collection.  Added item will be marked as "isNew".
+       * Add a new item to a collection. Added item is marked as "isNew".
        * @param {VuexTrackedInstance} state
        * @param {object} payload
        * @param {string} payload.collection Collection name
@@ -845,8 +847,12 @@ function createProcessingModule ({
             await saved()
           }
         } catch (err) {
-          UB.showErrorWindow(err)
-          throw new UB.UBAbortError(err)
+          if (typeof errorNotification === 'function') {
+            errorNotification(err)
+          } else {
+            UB.showErrorWindow(err)
+            throw new UB.UBAbortError(err)
+          }
         } finally {
           commit('LOADING', {
             isLoading: false,