+++ packages/org/org.js
@@ -67,7 +67,7 @@ function orgOnUserLogin () {
 
   if (!staffs.length) {
     // allow anonymous login only for SUPERUSER (before @unitybase/org@5.3 - a member of admin group)
-    if (ORG.checkOrgUnitRequired && !UBA_COMMON.isSuperUser()) {
+    if (ORG.checkOrgUnitRequired && !UBA_COMMON.isSuperUser() && !Session.uData.roleIDs.has(UBA_COMMON.ROLES.ADMIN.ID)) {
       throw new UB.UBAbort('<<<UserWithoutOrgEmployeeNotAllowed>>>')
     } else {
       // defined by ub model data.employeeShortFIO = ''