+++ packages/org/public/forms/org_staffunit-fm.vue
@@ -57,7 +57,10 @@
 const { Form, mapInstanceFields } = require('@unitybase/adminui-vue')
 module.exports.mount = cfg => {
   Form(cfg)
-    .processing()
+    .processing({
+      masterFieldList: ['parentID', 'professionID', 'staffUnitTypeID', 'professionExtID', 'code', 'name', 'fullName',
+        'description', 'subordinationLevel', 'isBoss', 'nameGen', 'nameDat', 'fullNameGen', 'fullNameDat']
+    })
     .validation()
     .mount()
 }
@@ -125,7 +128,7 @@ module.exports.default = {
         ...cfg,
         isModal: true,
         parentContext: {
-          stuffUnitID: this.ID
+          staffUnitID: this.ID
         }
       }
     },