+++ packages/ub/CHANGELOG.md
@@ -15,6 +15,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.7.9] - 2020-11-25
+## [5.7.8] - 2020-11-20
+### Added
+ - new `privateSettings` property for entity and attribute added to the JSON schema for entity metadata
+  
+## [5.7.7] - 2020-11-19
+### Fixed
+  - `UB.UBAbort` server side abort error now logged as `ERR` log level instead of `EXC` (as should be) for UBServer@5.18.20.
+  For UBServer < 5.18.20 nothing changed and UB.UBAbort will be logged as `EXC`.
+    
+
 ## [5.7.6] - 2020-11-15
 ## [5.7.5] - 2020-11-14
 ## [5.7.4] - 2020-11-12